fix(heat_pump_state): corrects conditional to set the new state and the function signature

+++ main/heat_pump_state.h
@@ -44,7 +44,7 @@ void heat_pump_state_destroy(StateData*);
     #include <time.h>
 
     // Private functions
-    void set_energy_consumption_state(StateData*, int, enum EnergyConsumptionState);
+    void set_energy_consumption_state(StateData*, unsigned long long, enum EnergyConsumptionState);
     char * match_state_from_tokens_object(int);
 
     #endif
